pyproject.toml
[project]
name = "corki"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "openai>=1.78.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sseclient-py>=1.8.0",
    "uvicorn>=0.34.2",
]
